updated utils, moved old parser back into action for non-x86 platforms
