+++ internal/transport/http1/transport.go
@@ -16,7 +16,7 @@ type Transport struct {
 
 func New(
 	request *http.Request,
-	keyBuff, valBuff, startLineBuff buffer.Buffer[byte],
+	keyBuff, valBuff, startLineBuff buffer.Buffer,
 	valuesPool pool.ObjectPool[[]string],
 	headersSettings settings.Headers,
 	respBuff, respFileBuff []byte,
